requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
colorful==0.5.4
cycler==0.10.0
Flask==1.1.2
Flask-SQLAlchemy==2.4.4
geographiclib==1.50
geopy==2.0.0
gunicorn==20.0.4
idna==2.10
importlib-metadata==4.5.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.2.0
MarkupSafe==1.1.1
matplotlib==3.3.1
numpy==1.19.1
Pillow==7.2.0
prettyprinter==0.18.0
Pygments==2.9.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.23
typing-extensions==3.10.0.0
urllib3==1.25.10
Werkzeug==1.0.1
zipp==3.4.1
